# backend/swagger.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Webhook Manager API",
    "description": "API documentation for Webhook Manager",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/wh/{urlPath}": {
      "get": {
        "parameters": [
          {
            "schema": { "type": "string" },
            "in": "path",
            "name": "urlPath",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "post": {
        "parameters": [
          {
            "schema": { "type": "string" },
            "in": "path",
            "name": "urlPath",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "put": {
        "parameters": [
          {
            "schema": { "type": "string" },
            "in": "path",
            "name": "urlPath",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "delete": {
        "parameters": [
          {
            "schema": { "type": "string" },
            "in": "path",
            "name": "urlPath",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      },
      "patch": {
        "parameters": [
          {
            "schema": { "type": "string" },
            "in": "path",
            "name": "urlPath",
            "required": true
          }
        ],
        "responses": { "200": { "description": "Default Response" } }
      }
    }
  }
}